# contracts/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Proyecta360 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/tasks": {
      "post": {
        "summary": "Create Task",
        "operationId": "create_task_api_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Task Api Tasks Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}": {
      "put": {
        "summary": "Update Task",
        "operationId": "update_task_api_tasks__task_id__put",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Update Task Api Tasks  Task Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Task",
        "operationId": "delete_task_api_tasks__task_id__delete",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Task Api Tasks  Task Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/indent": {
      "post": {
        "summary": "Indent Task",
        "operationId": "indent_task_api_tasks__task_id__indent_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Indent Task Api Tasks  Task Id  Indent Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/outdent": {
      "post": {
        "summary": "Outdent Task",
        "operationId": "outdent_task_api_tasks__task_id__outdent_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Outdent Task Api Tasks  Task Id  Outdent Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tasks/{task_id}/toggle": {
      "post": {
        "summary": "Toggle Task",
        "operationId": "toggle_task_api_tasks__task_id__toggle_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Toggle Task Api Tasks  Task Id  Toggle Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dependencies": {
      "post": {
        "summary": "Create Dependency",
        "operationId": "create_dependency_api_dependencies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DependencyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Create Dependency Api Dependencies Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/dependencies/{dependency_id}": {
      "delete": {
        "summary": "Delete Dependency",
        "operationId": "delete_dependency_api_dependencies__dependency_id__delete",
        "parameters": [
          {
            "name": "dependency_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dependency Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Dependency Api Dependencies  Dependency Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DependencyIn": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "predecessor_id": {
            "type": "integer",
            "title": "Predecessor Id"
          },
          "successor_id": {
            "type": "integer",
            "title": "Successor Id"
          },
          "dependency_type": {
            "type": "string",
            "title": "Dependency Type",
            "default": "FS"
          },
          "lag_days": {
            "type": "integer",
            "title": "Lag Days",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "predecessor_id",
          "successor_id"
        ],
        "title": "DependencyIn"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaskIn": {
        "properties": {
          "project_id": {
            "type": "integer",
            "title": "Project Id"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "component_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Component Id"
          },
          "title": {
            "type": "string",
            "maxLength": 220,
            "minLength": 1,
            "title": "Title"
          },
          "phase": {
            "type": "string",
            "title": "Phase",
            "default": ""
          },
          "task_type": {
            "type": "string",
            "title": "Task Type",
            "default": "task"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "duration_days": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Duration Days",
            "default": 1
          },
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress",
            "default": 0
          },
          "owner": {
            "type": "string",
            "title": "Owner",
            "default": ""
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "Pendiente"
          },
          "story_points": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Story Points",
            "default": 0
          },
          "budget": {
            "type": "number",
            "minimum": 0.0,
            "title": "Budget",
            "default": 0
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "order_index": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Order Index",
            "default": 0
          },
          "outline_level": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Outline Level",
            "default": 0
          },
          "is_expanded": {
            "type": "integer",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Is Expanded",
            "default": 1
          },
          "predecessor_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predecessor Id"
          },
          "dependency_type": {
            "type": "string",
            "title": "Dependency Type",
            "default": "FS"
          },
          "lag_days": {
            "type": "integer",
            "title": "Lag Days",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "title",
          "start_date"
        ],
        "title": "TaskIn"
      },
      "TaskUpdate": {
        "properties": {
          "parent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "component_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Component Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 220,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phase"
          },
          "task_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Type"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "duration_days": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Days"
          },
          "progress": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "story_points": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Story Points"
          },
          "budget": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "order_index": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Index"
          },
          "outline_level": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Outline Level"
          },
          "is_expanded": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Expanded"
          }
        },
        "type": "object",
        "title": "TaskUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}